backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
bcrypt==3.2.0
mysql-connector-python==8.0.26
PyMySQL==1.0.2
Flask-Caching==1.10.1